requirements.txt
ultralytics
pillow
streamlit